feat: replace UserSettingsStore with TauriSettingsStore and update related tests
